requirements.txt
Django==4.0.5
djoser==2.1.0
djangorestframework==3.13.1
Pillow==9.1.1